requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pinecone==7.3.0 
pypdf==4.2.0
google-genai==0.6.0
sentence-transformers==2.7.0
Jinja2==3.1.4

# LangChain components
langchain-community==0.3.27
langchain-google-genai==2.1.7
langchain-pinecone==0.2.8
langchain-core==0.3.68
langsmith==0.4.5